Added files for a misunderstood implementation
Assumed that the method would use Jacobi iteration to solve a given
linear system (for Ax=b, given A and b find x). Instead, a
implementation is needed that, given constant conditions on the boundary
of a 2D grid, find the steady state of the interior points and use a
Jacobi iterative solver to get those points.