1 node version of Jacobi Solver on Laplace Equation
Given a matrix with a border (b) of values, uses Jacobi iterative
solver approach to find the steady state of the interior points.
Points in the border aren't changed and has only be checked with
b=1.

Stencil such that a point in question is the average of the 4 points
around it.