Beginning of CPU and OpenACC code split
To do scaling comparisons, it would be better for the code to be
split into CPU and OpenACC versions so the comparison between 1
full CPU (36 ranks) can be compared to 1 GPU.

Also adds a routine to gather all MPI local matrices onto rank 0
and compare against results from running on 1 thread. NOTE: the
calls do to this comparison are currently commented out in the
main since it could cause the program to be memory hungry.